# api-spec.yaml
openapi: 3.1.0
info:
  title: BookWorm API
  version: 1.0.0
  description: Full-stack e-bookstore API — Next.js Route Handlers, PostgreSQL, JWT auth, Razorpay payment.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id:        { type: string, format: uuid }
        email:     { type: string, format: email }
        fullName:  { type: string }
        giftPoints:{ type: integer }
        role:      { type: string, enum: [customer, admin] }
    Book:
      type: object
      properties:
        id:              { type: string, format: uuid }
        title:           { type: string }
        author:          { type: string }
        price:           { type: number }
        format:          { type: string, enum: [paperback, hardcover, ebook] }
        language:        { type: string }
        cover_image_url: { type: string }
        description:     { type: string }
        author_bio:      { type: string }
        rating:          { type: number }
        sales_count:     { type: integer }
        delivery_days:   { type: integer }
        stock:           { type: integer }
        is_featured:     { type: boolean }
        category_name:   { type: string }
        category_slug:   { type: string }
        publisher_name:  { type: string }
        publisher_slug:  { type: string }
    CartItem:
      type: object
      properties:
        id:          { type: string, format: uuid }
        quantity:    { type: integer }
        book_id:     { type: string, format: uuid }
        title:       { type: string }
        author:      { type: string }
        price:       { type: number }
        format:      { type: string }
        cover_image_url: { type: string }
        delivery_days:   { type: integer }
...
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, email, password, confirmPassword]
              properties:
                fullName:        { type: string }
                email:           { type: string, format: email }
                password:        { type: string, minLength: 8 }
                confirmPassword: { type: string }
      responses:
        '201':
          description: Registered — auth_token cookie is set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Email already in use }
  /api/auth/login:
    post:
      summary: Login with email and password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Logged in — auth_token cookie is set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { description: Invalid credentials }
...
  /api/auth/me:
    get:
      summary: Get current authenticated user
      tags: [Auth]
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: Current user or null
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    oneOf:
                      - $ref: '#/components/schemas/User'
                      - type: 'null'

  # ── Categories ──────────────────────────────────────────────────────────────
...
  /api/books:
    get:
      summary: List and filter books
      tags: [Catalogue]
      parameters:
        - { name: category,  in: query, schema: { type: string }, description: Category slug }
        - { name: publisher, in: query, schema: { type: string }, description: Publisher slug }
        - { name: format,    in: query, schema: { type: string, enum: [paperback, hardcover, ebook] } }
        - { name: language,  in: query, schema: { type: string } }
        - { name: minPrice,  in: query, schema: { type: number } }
        - { name: maxPrice,  in: query, schema: { type: number } }
        - { name: sort,      in: query, schema: { type: string, enum: [relevance, price_asc, price_desc, bestselling, new_arrivals] } }
        - { name: q,         in: query, schema: { type: string }, description: Search query }
        - { name: page,      in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize,  in: query, schema: { type: integer, default: 16 } }
      responses:
        '200':
          description: Paginated book list
          content:
            application/json:
              schema:
                type: object
                properties:
                  books:    { type: array, items: { $ref: '#/components/schemas/Book' } }
                  total:    { type: integer }
                  page:     { type: integer }
                  pageSize: { type: integer }
...
  /api/cart:
    get:
      summary: Get cart items
      tags: [Cart]
      responses:
        '200':
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartItem'
    post:
      summary: Add item to cart
      tags: [Cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId]
              properties:
                bookId:   { type: string, format: uuid }
                quantity: { type: integer, default: 1 }
      responses:
        '201': { description: Added to cart }
        '404': { description: Book not found }
  /api/cart/{itemId}:
    patch:
      summary: Update cart item quantity
      tags: [Cart]
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 1 }
      responses:
        '200': { description: Updated }
    delete:
      summary: Remove item from cart
      tags: [Cart]
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Removed }

  # ── Wishlist ─────────────────────────────────────────────────────────────────
...
  /api/checkout/cancel:
    post:
      summary: Cancel an order within 48 hours
      tags: [Checkout]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId]
              properties:
                orderId: { type: string, format: uuid }
      responses:
        '200': { description: Order cancelled }
        '400': { description: Cancellation window passed or already cancelled }
        '404': { description: Order not found }